// src/app/core/enums/VariableType.ts
export enum VariableType {
    PrinterVariable = 'PrinterVariable',
    MaterialVariable = 'MaterialVariable',
    PricingModelVariable = 'PricingModelVariable',
    ProcessVariable = 'ProcessVariable',
    FinishesVariable = 'FinishesVariable',
    OrderVariable = 'OrderVariable'
  }


  export const variableTypeDisplayNames = {
    PrinterVariable: 'Printer variable',
    MaterialVariable: 'Material variable',
    PricingModelVariable: 'Pricing model variable',
    ProcessVariable: 'Process variable',
    FinishesVariable: 'Finishes variable',
    OrderVariable: 'Order variable'
  };